+++ CSK_Module_UserManagement/scripts/Configuration/UserManagement/helper/checkAPIs.lua
@@ -1,0 +1,33 @@
+---@diagnostic disable: undefined-global, redundant-parameter, missing-parameter
+
+-- Load all relevant APIs for this module
+--**************************************************************************
+
+local availableAPIs = {}
+
+local function loadAPIs()
+  CSK_UserManagement = require 'API.CSK_UserManagement'
+
+  Cipher = {}
+  Cipher.AES = require 'API.Cipher.AES'
+  Container = require 'API.Container'
+  Engine = require 'API.Engine'
+  Log = require 'API.Log'
+  Log.Handler = require 'API.Log.Handler'
+  Log.SharedLogger = require 'API.Log.SharedLogger'
+  Object = require 'API.Object'
+  Timer = require 'API.Timer'
+
+  -- Check if related CSK modules are available to be used
+  local appList = Engine.listApps()
+  for i = 1, #appList do
+    if appList[i] == 'CSK_Module_PersistentData' then
+      CSK_PersistentData = require 'API.CSK_PersistentData'
+    end
+  end
+end
+
+availableAPIs.default = xpcall(loadAPIs, debug.traceback) -- TRUE if all default APIs were loaded correctly
+
+return availableAPIs
+--**************************************************************************